Fix for old version of node
